Upon victory, users can submit name.
Created a form that renders on game over.

Necessary conditions that are stored locally are reset on game over (time and hits).

Submission form is styled and replaces the entirety of the game page when active.

Correction made to firebase.js where Firestore was written as FireStore.

FinalScoreContext and GameoverContext created.