Fixed a refresh exploit on the submit page.
If the user refreshed the page during the score submit screen, they could play the game at 0 time without the clock moving forward. Now the page will refresh to the home page.

The final check used to leave a blue square behind that was out of place. A conditional prevents that from happening now.

The game used to check for 9 Straw Hats, now it checks for 10, but a bug still exists where it takes one more correct guess to initiate the submission page.